Make the acctest helper script a little easier to use
The dns/acceptance.sh script starts up a bind server in docker to provide
a target to run the acceptance tests against.

This is a couple minor adjustments to that script to make it easier to
use:

- Run the local test DNS server on a non-standard port, to avoid conflicts
  with a real nameserver running on localhost:53

- Clean up the docker containers even if the tests fail, to avoid the
  need for manual cleanup in that case.

In order to implement this it was necessary to make the update port number
configurable in the environment, similar to all of the other settings.