Update TestAccDataDnsTxtRecordSet_Basic for changes to DNS zone
Several of our tests unfortunately still depend on real domains, such as
terraform.io in this case, and so they end up failing when legitimate
changes are made.

One day we should move this over to an explicit test domain, but for now
we'll just update the test to match the latest results in the real domain.