Bump github.com/miekg/dns dependency
Updating provides the ability to plug in different TSIG implementations
which is needed for #30. The current version was quite old too.

This commit is the result of running the following:

* go get -u github.com/miekg/dns@master
* go mod tidy
* go mod vendor

After updating, a small change was needed to the code as the
dns.ErrTruncated error has been dropped from the library. I've updated
the code to instead check the Truncated flag on the response which is
apparently the preferred way to do it now. There should be no overall
change in behaviour.