Update acceptance tests
Switch to using our own Dockerfile. This creates three container images:

* Kerberos KDC. This creates a new realm and two keytab files; one for
  BIND to use and another for a test user.
* BIND DNS server. This has Kerberos configured but can be operated
  without.
* User keytab. This exists purely to extract the keytab file out of the
  KDC container.

The containers need to be created in sync with each other as the keytab
files need to match up.

The acceptance tests now build the containers from scratch each time,
although the layers should be cached so repeat builds will be fast.

The tests then run as before using the BIND container with a different
/etc/named.conf volume mounted into place with either no TSIG
configured, an HMAC-SHA256 TSIG key, or using Kerberos/GSS-TSIG. In the
last case the KDC container is also created.

For now, no Kerberos-enabled tests are actually run.

The Travis build environment needed to be bumped from Trusty to Focal to
get a new enough version of Docker to be able to use the buildx plugin,
which is also installed.