Update documentation to latest Terraform syntax
This updates the documentation to Terraform version 0.11+ syntax, which
allows constant expressions to be provided without the interpolation
syntax

Current documentation snippet:

```
output "google_addrs" {
  value = "${join(",", data.dns_a_record_set.google.addrs)}"
}
```

Terraform output:

```
Warning: Interpolation-only expressions are deprecated

  on test.tf line 71, in output "dustycloth_addrs":
  71:   value = "${join(",", data.dns_a_record_set.google.addrs)}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.
```